Conflicts:
	tests/sql/src/main/java/sql/dmlStatements/TradeBuyOrdersDMLStmt.java

+++ gemfirexd/core/src/main/java/com/pivotal/gemfirexd/internal/engine/db/FabricDatabase.java
@@ -440,11 +440,12 @@ public final class FabricDatabase implements ModuleControl,
       notifyRunning();
       return;
     }
+    final GemFireCacheImpl cache = GemFireCacheImpl.getExisting();
+    final LogWriter logger = cache.getLogger();
+    final EmbedConnection embedConn = (EmbedConnection)conn;
 
+    List<GfxdSystemProcedureMessage> postMsgs = null;
     try {
-      final EmbedConnection embedConn = (EmbedConnection)conn;
-      final GemFireCacheImpl cache = GemFireCacheImpl.getExisting();
-      final LogWriter logger = cache.getLogger();
       final LanguageConnectionContext lcc = embedConn.getLanguageConnection();
       final GemFireTransaction tc = (GemFireTransaction)lcc
           .getTransactionExecute();
@@ -488,7 +489,7 @@ public final class FabricDatabase implements ModuleControl,
         else {
           this.memStore.getDDLStmtQueue().initializeQueue(this.dd);
         }
-        postCreateDDLReplay(embedConn, bootProps, lcc, tc, logger);
+        postMsgs =  postCreateDDLReplay(embedConn, bootProps, lcc, tc, logger);
       } finally {
         if (ddReadLockAcquired) {
           this.dd.unlockAfterReading(null);
@@ -508,7 +509,7 @@ public final class FabricDatabase implements ModuleControl,
       initializeCatalog();
     } catch (Throwable t) {
       try {
-        LogWriter logger = Misc.getCacheLogWriter();
+
         if (logger != null) {
           logger.warning("got throwable: " + t.getMessage() + " calling shut down", t);
         }
@@ -539,6 +540,30 @@ public final class FabricDatabase implements ModuleControl,
       throw StandardException.newException(SQLState.BOOT_DATABASE_FAILED, t,
           Attribute.GFXD_DBNAME);
     }
+
+
+    for (GfxdSystemProcedureMessage msg : postMsgs) {
+      if (msg.getSysProcMethod().isOffHeapMethod()
+        && this.memStore.getGemFireCache().getOffHeapStore() == null) {
+        if (logger.severeEnabled()) {
+          logger.severe("FabricDatabase: aborted initial replay "
+            + "for message " + msg + " method "
+            + msg.getSysProcMethod().name());
+        }
+        continue;
+      }
+      try {
+        msg.execute();
+      } catch (Exception ex) {
+        if (logger.severeEnabled()) {
+          logger.severe("FabricDatabase: failed initial replay "
+            + "for message " + msg + " due to exception", ex);
+        }
+        throw StandardException.newException(SQLState.BOOT_DATABASE_FAILED, ex,
+          Attribute.GFXD_DBNAME);
+      }
+    }
+
   }
 
   private Region createSnappySpecificGlobalCmdRegion(boolean isLead) throws IOException, ClassNotFoundException {
@@ -658,7 +683,7 @@ public final class FabricDatabase implements ModuleControl,
 
     try {
       lcc.getDataDictionary().lockForReading(tc);
-      hiveDBTablesMap = externalCatalog.getAllStoreTablesInCatalog();
+      hiveDBTablesMap = externalCatalog.getAllStoreTablesInCatalogUppercase();
       gfDBTablesMap = getAllGFXDTables();
     } finally {
       lcc.getDataDictionary().unlockAfterReading(tc);
@@ -889,7 +914,7 @@ public final class FabricDatabase implements ModuleControl,
    * Replays the initial DDL received by GII from other nodes or recovered from
    * disc.
    */
-  private void postCreateDDLReplay(final EmbedConnection embedConn,
+  private List<GfxdSystemProcedureMessage> postCreateDDLReplay(final EmbedConnection embedConn,
       final Properties bootProps, final LanguageConnectionContext lcc,
       final GemFireTransaction tc, final LogWriter logger) throws Exception {
 
@@ -970,7 +995,7 @@ public final class FabricDatabase implements ModuleControl,
     final LinkedHashSet<GemFireContainer> uninitializedTables =
         new LinkedHashSet<GemFireContainer>();
     final Statement stmt = embedConn.createStatement();
-
+    List<GfxdSystemProcedureMessage> postMessages = new ArrayList<>();
     try {
       // commenting out for snap-585
       /*
@@ -1067,6 +1092,7 @@ public final class FabricDatabase implements ModuleControl,
             .getPreprocessedDDLQueue(currentQueue, skipRegionInit,
                 lastCurrentSchema, pre11TableSchemaVer, traceConflation);
 
+
         for (GfxdDDLQueueEntry entry : preprocessedQueue) {
           qEntry = entry;
           Object qVal = qEntry.getValue();
@@ -1083,6 +1109,10 @@ public final class FabricDatabase implements ModuleControl,
           if (qVal instanceof GfxdSystemProcedureMessage) {
             final GfxdSystemProcedureMessage msg =
                 (GfxdSystemProcedureMessage)qVal;
+            if (msg.postprocess()) {
+              postMessages.add(msg);
+              continue;
+            }
             if (msg.getSysProcMethod().isOffHeapMethod()
                 && this.memStore.getGemFireCache().getOffHeapStore() == null) {
               if (logger.severeEnabled()) {
@@ -1500,12 +1530,13 @@ public final class FabricDatabase implements ModuleControl,
       stmt.close();
       // Setting this to false so that the waiting compactor thread finishes
       this.memStore.setInitialDDLReplayInProgress(false);
-    }
+      // restore the original schema if required
+      if (!ArrayUtils.objectEquals(initSchema, lcc.getCurrentSchemaName())) {
+        FabricDatabase.setupDefaultSchema(dd, lcc, tc, initSchema, true);
+      }
 
-    // restore the original schema if required
-    if (!ArrayUtils.objectEquals(initSchema, lcc.getCurrentSchemaName())) {
-      FabricDatabase.setupDefaultSchema(dd, lcc, tc, initSchema, true);
     }
+    return  postMessages;
   }
 
   private void checkRecoveredIndex(ArrayList<GemFireContainer> uninitializedContainers,
@@ -1863,8 +1894,12 @@ public final class FabricDatabase implements ModuleControl,
     // Clean up GemFireXD MBeans if management was not disabled
     //GfxdManagementService.handleEvent(GfxdResourceEvent.FABRIC_DB__STOP, this.memStore);
     active = false;
-    tempDir.deleteAll();
-    tempDir = null;
+    final DirFile tempDir = this.tempDir;
+    if (tempDir != null) {
+      // noinspection ResultOfMethodCallIgnored
+      tempDir.deleteAll();
+      this.tempDir = null;
+    }
     runtimeStatisticsOn = false;
   }
 
@@ -2545,8 +2580,8 @@ public final class FabricDatabase implements ModuleControl,
                     .nextInt(Integer.MAX_VALUE);
                 final DirFile df = new DirFile(tDir, TEMP_DIR_PREFIX
                     + Integer.toString(rl) + ".d");
-                df.deleteOnExit();
                 if (df.mkdirs()) {
+                  df.deleteOnExit();
                   assert df.canWrite();
                   return df;
                 }
